requirements.txt
amqp==5.1.1
anyio==4.0.0
arabic-reshaper==3.0.0
asgiref==3.5.2
asn1crypto==1.5.1
async-timeout==4.0.3
autopep8==2.0.4
backports.zoneinfo==0.2.1
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
cryptography==41.0.4
cssselect2==0.7.0
Django==4.0.4
django-appconf==1.0.5
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-ckeditor==6.4.2
django-environ==0.9.0
django-imagekit==4.1.0
django-js-asset==2.0.0
django-phonenumber-field==7.1.0
django-timezone-field==6.0.1
exceptiongroup==1.1.3
flake8==6.1.0
gunicorn==21.2.0
h11==0.14.0
html5lib==1.1
httptools==0.6.0
idna==3.4
kombu==5.3.2
lxml==4.9.3
mccabe==0.7.0
msgpack==1.0.7
oscrypto==1.3.0
packaging==23.2
phonenumbers==8.13.20
pilkit==2.0
Pillow==9.1.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.1
pycodestyle==2.11.0
pycparser==2.21
pyflakes==3.1.0
pyHanko==0.20.1
pyhanko-certvalidator==0.24.1
pyotp==2.9.0
pypdf==3.16.2
pypi==2.1
pypng==0.20220715.0
python-bidi==0.4.2
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-http-client==3.3.7
python-resize-image==1.1.20
python-slugify==8.0.1
PyYAML==6.0.1
qrcode==7.4.2
razorpay==1.4.1
redis==5.0.1
reportlab==3.6.13
requests==2.31.0
sendgrid==6.8.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
sqlparse==0.4.2
starkbank-ecdsa==2.2.0
svglib==1.5.1
text-unidecode==1.3
tinycss2==1.2.1
tomli==2.0.1
typing-extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
uritools==4.0.2
urllib3==2.0.4
uvicorn==0.23.2
uvloop==0.18.0
vine==5.0.0
watchfiles==0.21.0
wcwidth==0.2.8
webencodings==0.5.1
websockets==11.0.3
whitenoise==6.6.0
xhtml2pdf==0.2.11
